On Windows long is not uint64_t

+++ x86/x86.cpp
@@ -104,7 +104,7 @@ void X86Assembler::renderInstruction(const RDRendererParams* srp)
             case ZYDIS_TOKEN_ADDRESS_ABS:
             case ZYDIS_TOKEN_ADDRESS_REL:
             case ZYDIS_TOKEN_IMMEDIATE:
-                RDRenderer_Unsigned(srp->renderer, std::stoul(tokenvalue, nullptr, 16));
+                RDRenderer_Unsigned(srp->renderer, static_cast<u64>(std::stoull(tokenvalue, nullptr, 16)));
                 break;
 
             case ZYDIS_TOKEN_MNEMONIC:
